Lib: move to root directory
