+++ lib/store.js
