+++ lib/tests/crypto.test.js
