+++ lib/tests/stubs.js
